bugfixes and viz refactor
